move tests into test_utils file
